The version 4.0.0 releases the following issues:
- Improve abstraction of C4SharpDirectory.cs class
- Enable Nullable Warning
- Improve abstraction of PlantumlResources.cs class
- Improve PlantumlSession.cs responsibility
- Improve Abstraction of Plantuml module
