+++ .github/workflows/dotnet.yml
@@ -7,7 +7,7 @@ on:
     branches: [ main ]
     
 env:
-  VERSION: 3.2.${{ github.run_number }}
+  VERSION: 4.0.${{ github.run_number }}
   NUGET_INDEX: https://api.nuget.org/v3/index.json
   BUILD_TYPE: Release 
 
@@ -34,7 +34,7 @@ jobs:
       run: dotnet build --no-restore --configuration Release -p:PackageVersion=${{env.VERSION}}
       
     - name: Test
-      run: dotnet test ./tests/**/*.UnitTests.csproj --no-restore
+      run: dotnet test ./tests/**/*.*Tests.csproj --no-restore
 
     - name: Publish Artifact
       run: |
