+++ samples/C4Sharp.Sample.Orders/Program.cs
@@ -1,11 +1,11 @@
 ﻿using C4Sharp.Diagrams;
-using C4Sharp.Models.Plantuml;
+using C4Sharp.Models.Plantuml.IO;
 
 namespace C4Sharp.Sample.Orders
 {
     internal static class Program
     {
-        private static void Main(string[] args)
+        private static void Main()
         {
             var diagrams = new Diagram[]
             {
