+++ src/C4Sharp/Diagrams/Diagram.cs
@@ -15,7 +15,7 @@ namespace C4Sharp.Diagrams
         public bool LayoutWithLegend { get; set; }
         public bool ShowLegend { get; set; }
         public bool LayoutAsSketch { get; set; }
-        public string Title { get; set; }
+        public string? Title { get; set; }
         public DiagramLayout FlowVisualization { get; set; }
         public Structure[] Structures { get; set; }
         public Relationship[] Relationships { get; set; }
