+++ src/C4Sharp/Extensions/ResourceMethods.cs
@@ -2,7 +2,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Reflection;
-using C4Sharp.Models.Plantuml;
+using C4Sharp.Models.Plantuml.IO;
 
 namespace C4Sharp.Extensions
 {
@@ -13,23 +13,16 @@ namespace C4Sharp.Extensions
     internal static class ResourceMethods
     {
         /// <summary>
-        /// Get Stream from plantuml.jar file
-        /// </summary>
-        /// <returns>Stream</returns>
-        /// <exception cref="PlantumlException"></exception>
-        public static Stream GetPlantumlResource() => GetResourceByName("C4Sharp.Resources.plantuml.jar");
-
-        /// <summary>
         /// Get resource string content from resource file
         /// </summary>
         /// <param name="name">file name</param>
         /// <returns>resource content</returns>
         /// <exception cref="PlantumlException"></exception>
-        public static string GetResource(string name)
+        public static string GetResourceContent(string name)
         {
             try
             {
-                using var stream = GetResourceByName($"C4Sharp.Resources.{name}");
+                using var stream = GetResourceStream(name);
                 using var reader = new StreamReader(stream ?? throw new InvalidOperationException());
                 return reader.ReadToEnd();
             }
@@ -45,13 +38,13 @@ namespace C4Sharp.Extensions
         /// <param name="name"></param>
         /// <returns></returns>
         /// <exception cref="PlantumlException"></exception>
-        private static Stream GetResourceByName(string name)
+        public static Stream? GetResourceStream(string name)
         {
             try
             {
                 return Assembly
                     .GetExecutingAssembly()
-                    .GetManifestResourceStream(name);
+                    .GetManifestResourceStream($"C4Sharp.Resources.{name}");
             }
             catch (Exception e)
             {
