+++ tests/C4Sharp.IntegratedTests/GettingStreamDiagramTests.cs
@@ -1,6 +1,5 @@
-using C4Sharp.Diagrams;
 using C4Sharp.IntegratedTests.Stubs.Diagrams;
-using C4Sharp.Models.Plantuml;
+using C4Sharp.Models.Plantuml.IO;
 using FluentAssertions;
 using Xunit;
